projects/hcp/backend/pyproject.toml
[project]
name = "hcp-s3-api"
version = "0.1.0"
description = "Unified API gateway for Hitachi Content Platform (MAPI, S3, Metadata Query)"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aioboto3>=15.0",
    "fastapi>=0.133.1",
    "httpx>=0.28.1",
    "opentelemetry-api>=1.29.0",
    "opentelemetry-sdk>=1.29.0",
    "opentelemetry-instrumentation-fastapi>=0.50b0",
    "opentelemetry-instrumentation-httpx>=0.50b0",
    "opentelemetry-exporter-otlp>=1.29.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.11.0",
    "python-multipart>=0.0.22",
    "py-key-value-aio[redis,memory]>=0.2",
    "pyarrow>=17.0",
    "tenacity>=9.1.4",
]

[project.optional-dependencies]
serve = [
    "uvicorn[standard]>=0.41.0",
    "gunicorn>=23.0",
]
lance = [
    "lancedb>=0.29",
    "pylance>=2.0.1",
]

[dependency-groups]
dev = [
    "moto[s3]>=5.0",
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
    "pytest-cov>=6.0",
    "httpx>=0.28.1",
    "respx>=0.21",
    "ruff>=0.11",
    "ty>=0.0.1a7",
]

docs = [
    "zensical>=0.0.24",
    "mkdocstrings-python>=2.0.3",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
markers = [
    "redis: tests requiring a real Redis connection",
    "moto: integration tests using moto S3 mock",
    "minio: integration tests requiring a real MinIO service",
]
